Controle de uso. Issue #4

+++ PyRoles.py
@@ -34,17 +34,21 @@ def send_welcome(message):
 # salvando a foto no sistema e fazendo o upload para o flickr 
 @bot.message_handler(content_types=['photo'])
 def get_doc(message):
-    bot.send_chat_action(message.chat.id, 'upload_photo')
-    raw = message.photo[-1].file_id
-    path = raw+".jpg"
-    file_info = bot.get_file(raw)
-    downloaded_file = bot.download_file(file_info.file_path)
-    with open(path, 'wb') as new_file:
-        new_file.write(downloaded_file)
-    bot.reply_to(message, "A sua fotografia agora faz parte do <a href='https://www.flickr.com/photos/160228175@N08/'>álbum PyRolês</a> ! \nObrigada por fazer essa comunidade ser tão maravilhosa!💛💙", parse_mode="HTML", disable_web_page_preview=True)
-    flickr.upload(filename=path, title='PyBR14', description='Python Brasil [14]')
+    if str(message.chat.id) not in str(config['TGBOT']['ALLOWED']):
+        bot.reply_to(message, "Não permitido.")
+    else:
+        bot.send_chat_action(message.chat.id, 'upload_photo')
+        raw = message.photo[-1].file_id
+        path = raw+".jpg"
+        file_info = bot.get_file(raw)
+        downloaded_file = bot.download_file(file_info.file_path)
+        with open(path, 'wb') as new_file:
+            new_file.write(downloaded_file)
+        bot.reply_to(message, "A sua fotografia agora faz parte do <a href='https://www.flickr.com/photos/160228175@N08/'>álbum PyRolês</a> ! \nObrigada por fazer essa comunidade ser tão maravilhosa!💛💙", parse_mode="HTML", disable_web_page_preview=True)
+        flickr.upload(filename=path, title='PyBR14', description='Python Brasil [14]')
 
 # apaga a foto do servidor 
-    os.remove(path)
+        os.remove(path)
  
 bot.polling()
+
